+++ .storybook/preview.js
@@ -38,8 +38,8 @@ const theme = createTheme({
         root: ({ ownerState }) => ({
           ...(ownerState.variant === 'contained' &&
             ownerState.color === 'secondary' && {
-            boxShadow: '0px 0px 10px #f8c608',
-          }),
+              boxShadow: '0px 0px 10px #f8c608',
+            }),
         }),
       },
     },
@@ -96,4 +96,5 @@ export const parameters = {
       order: ['Introduction', 'Colors', 'Typography'],
     },
   },
+  layout: 'centered',
 };
