+++ package.json
@@ -78,7 +78,6 @@
     "@mui/material": "^5.6.2",
     "@mui/styles": "^5.8.0",
     "i18next": "^21.8.11",
-    "plausible-tracker": "^0.3.8",
     "react-string-replace": "^1.1.0",
     "reactjs-popup": "^2.0.5"
   },
