+++ src/components/SourceItem/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './SourceItem';
