+++ src/components/SupportIndicator/SupportIndicator.scss
@@ -1,6 +1,6 @@
 .SupportIndicatorContainer {
   margin-top: 0.5em;
-  margin-bottom: 0.5em;
+  margin-bottom: 0;
 }
 
 $red-common: #ff8761;
