+++ src/components/index.ts
@@ -15,3 +15,4 @@ export { default as SupportIcon } from './SupportIcon';
 export { default as Dropdown } from './Dropdown';
 export { default as ToggleButton } from './ToggleButton';
 export { default as SupportIndicator } from './SupportIndicator';
+export { default as SourceItem } from './SourceItem';
